fansite/requirements.txt
Django
django-allauth
django-filter
python-dotenv
requests
gunicorn
Pillow
pydantic
psycopg2-binary
django-ckeditor
redis
celery
django-crispy-forms
crispy-bootstrap4
